parmatch: factorize [satisfiable] and [satisfiables] (REVIEW REQUIRED)
The tests still pass so I expect that there was indeed no semantic change as a
result of this factorization. [1]
However I might have have *unreasonably* degraded performances in the simple
case. Please have a look.

Also, I'd like to rename [satisfiables] to something more meaningful,
[list_satisfying_vectors] is only partially convincing, suggestions welcomed.

Note:
[1]: in my first implementation I had forgotten to guard one of the calls to
[set_args] with the [if only_care_about_existence] check. As a result "make
world" succeeded, so did the tests typing-warnings/, but "make world.opt" failed
with a fatal error in [read_args].
Guarding none of the calls to [set_args] or all of them does fix the problem.
I chose to guard all of them, but that might actually have a negative impact on
performance in general.