Represent ancestor variables more directly
Previously, some of the translation of method calls on ancestor
variables (bindings from `inherit` statements) was done in typecore.ml
by creating fake typedtree expressions. This has been moved into
translcore.ml which is simpler. We also pass the types of inherited
methods directly as part of the ancestor variable's kind, which
eleminates the need for the dummy `selfpat-n` variables.