matching: fix a very old bug in Default_environment.specialize_matrix
The bug was inserted in 842eb76ef7f5f3f85ce4fac22acf201fb237a238 in
2002. It breaks specialization of default matrices, so in theory it
should not lead to wrong-code production, but because it can create
default matrices of distinct arity in the same environment it should
be able to observe assertion failures in code hitting the bug.

I found the bug while upgrading the code to use better types. The
input of specialize_matrix has non_empty rows, but the output may be
empty -- and this code becomes ill-typed when input and output have
distinct types. Yay typing.